+++ subsys/bluetooth/host/gatt_internal.h
@@ -14,9 +14,6 @@
 #include <zephyr/bluetooth/addr.h>
 #include <zephyr/bluetooth/conn.h>
 
-#define BT_GATT_CENTRAL_ADDR_RES_NOT_SUPP	0
-#define BT_GATT_CENTRAL_ADDR_RES_SUPP		1
-
 #define BT_GATT_PERM_READ_MASK			(BT_GATT_PERM_READ | \
 						BT_GATT_PERM_READ_ENCRYPT | \
 						BT_GATT_PERM_READ_AUTHEN | \
